Refactor channel creation to ChannelsRegistry
