+++ things/src/main/java/com/example/androidthings/lantern/ChannelsRegistry.kt
@@ -1,9 +1,11 @@
 package com.example.androidthings.lantern
 
 import android.net.Uri
+import android.os.Bundle
 import com.example.androidthings.lantern.channels.*
 import com.example.androidthings.lantern.channels.nowplaying.NowPlayingChannel
 import com.example.androidthings.lantern.channels.spaceporthole.SpacePortholeChannel
+import com.example.androidthings.lantern.shared.ChannelConfiguration
 import com.example.androidthings.lantern.shared.ChannelInfo
 
 
@@ -85,4 +87,31 @@ object ChannelsRegistry {
     )
 
     val channelsInfo = channelsWithInfo.map { it.second }
+
+    fun newChannelForConfig(config: ChannelConfiguration): Channel {
+        val args = Bundle()
+        args.putParcelable(Channel.ARG_CONFIG, config)
+        var channel: Channel? = null
+        var rotationDisabled = false
+
+        for ((channelConstructor, info) in ChannelsRegistry.channelsWithInfo) {
+            if (info.id == config.type) {
+                channel = channelConstructor()
+                rotationDisabled = info.rotationDisabled
+                break
+            }
+        }
+
+        if (rotationDisabled) {
+            args.putBoolean(Channel.ARG_ROTATION_DISABLED, rotationDisabled)
+        }
+
+        if (channel == null) {
+            channel = ErrorChannel()
+            args.putString(ErrorChannel.ARG_MESSAGE, "Unknown channel type '${config.type}'")
+        }
+
+        channel.arguments = args
+        return channel
+    }
 }
